+++ Tests/SuperwallKitTests/Paywall/Presentation/Internal Presentation/Operators/WaitToPresentOperatorTests.swift
@@ -103,8 +103,9 @@ final class WaitToPresentTests: XCTestCase {
 
   func test_waitToPresent_hasIdentity_inactiveStatus_hasConfig_noInternet() {
     let expectation = expectation(description: "Got identity")
+    expectation.isInverted = true
 
-    let unknownSubscriptionPublisher = CurrentValueSubject<SubscriptionStatus, Never>(SubscriptionStatus.inactive)
+    let inactiveSubscriptionPublisher = CurrentValueSubject<SubscriptionStatus, Never>(SubscriptionStatus.inactive)
       .eraseToAnyPublisher()
 
     dependencyContainer.configManager.config = .stub()
@@ -116,13 +117,75 @@ final class WaitToPresentTests: XCTestCase {
       type: .getPaywallViewController(.stub()),
       hasInternetOverride: false
     )
-    .setting(\.flags.subscriptionStatus, to: unknownSubscriptionPublisher)
+    .setting(\.flags.subscriptionStatus, to: inactiveSubscriptionPublisher)
+
+    Task {
+      do {
+        try await Superwall.shared.waitToPresent(request, dependencyContainer: dependencyContainer)
+      } catch {
+        expectation.fulfill()
+      }
+    }
+
+    identityManager.didSetIdentity()
+
+    wait(for: [expectation], timeout: 0.1)
+  }
+
+  func test_waitToPresent_hasIdentity_inactiveStatus_noConfig_noInternet() {
+    let expectation = expectation(description: "Got identity")
+
+    let inactiveSubscriptionPublisher = CurrentValueSubject<SubscriptionStatus, Never>(SubscriptionStatus.inactive)
+      .eraseToAnyPublisher()
+
+    dependencyContainer.configManager.config = nil
+    let request = dependencyContainer.makePresentationRequest(
+      .explicitTrigger(.stub()),
+      paywallOverrides: nil,
+      isDebuggerLaunched: false,
+      isPaywallPresented: false,
+      type: .getPaywallViewController(.stub()),
+      hasInternetOverride: false
+    )
+    .setting(\.flags.subscriptionStatus, to: inactiveSubscriptionPublisher)
 
     Task {
       do {
         try await Superwall.shared.waitToPresent(request, dependencyContainer: dependencyContainer)
       } catch let error as PresentationPipelineError {
         if case .noInternet = error {
+          expectation.fulfill()
+        }
+      }
+    }
+
+    identityManager.didSetIdentity()
+
+    wait(for: [expectation], timeout: 0.1)
+  }
+
+  func test_waitToPresent_hasIdentity_activeStatus_noConfig_noInternet() {
+    let expectation = expectation(description: "Got identity")
+
+    let inactiveSubscriptionPublisher = CurrentValueSubject<SubscriptionStatus, Never>(SubscriptionStatus.active)
+      .eraseToAnyPublisher()
+
+    dependencyContainer.configManager.config = nil
+    let request = dependencyContainer.makePresentationRequest(
+      .explicitTrigger(.stub()),
+      paywallOverrides: nil,
+      isDebuggerLaunched: false,
+      isPaywallPresented: false,
+      type: .getPaywallViewController(.stub()),
+      hasInternetOverride: false
+    )
+    .setting(\.flags.subscriptionStatus, to: inactiveSubscriptionPublisher)
+
+    Task {
+      do {
+        try await Superwall.shared.waitToPresent(request, dependencyContainer: dependencyContainer)
+      } catch let error as PresentationPipelineError {
+        if case .userIsSubscribed = error {
           expectation.fulfill()
         }
       }