+++ CHANGELOG.md
@@ -21,7 +21,7 @@ The changelog for `SuperwallKit`. Also see the [releases](https://github.com/sup
 
 - Fixes issue where a crash would occur if storage was full and a persistent container couldn't be created.
 - Fixes thread safety issue when using a lazy variable to retrieve products.
-- If the internet is offline when trying to present a paywall, it now throws a presentationError. If the internet reconnects future paywalls will show.
+- If the internet is offline when trying to present a paywall, the paywall configuration hasn't been retrieved, and the user is not subscribed, it now throws a presentationError. If the internet reconnects future paywalls will show.
 - Fixes retry logic for requests.
 
 ## 3.0.0-rc.6